requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
openai==1.35.3
python-dotenv==1.0.1
jinja2==3.1.4
python-multipart==0.0.9
pydantic==1.10.13